api: extract requestGeo and getKeyData to shared lib
- lib/api/requestGeo: getClientIp, hashIp, getLocationFromIP
- lib/keyHashing: getKeyData for meta.key extraction
- Routes + rateLimit use shared utils
